Introducing unified module that includes WSApp (was WS) and WSClient.
