+++ ws_sdk/__init__.py
@@ -1,12 +1,14 @@
 from logging import getLogger
 
-from ws_sdk.web import WS
+from ws_sdk.app import WSApp
 from ws_sdk.client import WSClient
+from ws_sdk.web import WS
 
 logger = getLogger(__name__)
 
 __all__ = (
     "WS",
+    "WSApp",
     "WSClient",
     "ws_constants",
     "ws_utilities"
