+++ ws_sdk/client.py
@@ -27,7 +27,8 @@ class WSClient:
                  java_bin: str = JAVA_BIN,
                  ua_path: str = f"c:/tmp/ws-{__tool_name__}" if sys.platform == "win32" else f"/tmp/{__tool_name__}",
                  skip_ua_download: bool = False,
-                 tool_details: tuple = ("ps-sdk", __version__)
+                 tool_details: tuple = ("ps-sdk", __version__),
+                 **kwargs
                  ):
         if token_type is ORGANIZATION:
             self.ua_path = ua_path
@@ -36,11 +37,12 @@ class WSClient:
             self.ua_jar_f_with_path = os.path.join(ua_path, UA_JAR_F_N)
             # UA configuration
             self.ua_conf = ws_utilities.WsConfiguration()
+            self.ua_conf.projectPerFolder = False
             self.ua_conf.apiKey = token
             self.ua_conf.userKey = user_key
             self.ws_url = f"{ws_utilities.get_full_ws_url(url)}"
             self.java_bin = java_bin if bool(java_bin) else JAVA_BIN
-            self.ua_conf.wss_url = self.get_client_api_url(self.ws_url)
+            self.ua_conf.wss_url = WSClient.get_client_api_url(self.ws_url)
             self.ua_conf.log_files_path = self.ua_path
             self.ua_conf.whiteSourceFolderPath = self.ua_path
             self.ua_conf.checkPolicies = False
